bench/pyproject.toml
[project]
name = "mrmrs-bench"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "mrmrs",
    "mrmr_selection",
    "scikit-learn",
    "numpy",
    "pandas",
]

[tool.uv.sources]
mrmrs = { workspace = true }
